Add Stale Issue/PR closer (#1306)
There are a lot of stale issues and PRs from folks that don't seem involved anymore.  It's possible some are still relevant, so I set a pretty generous grace period:
Mark anything older than a year as stale, and close after 30 days of marking as stale.

Seeking feedback on thresholds and verbiage.